+++ lib/reach/smell/helpers.ex
@@ -25,6 +25,19 @@ defmodule Reach.Smell.Helpers do
     end
   end
 
+  def ast_modules_in_file(ast) do
+    {_ast, modules} =
+      Macro.prewalk(ast, [], fn
+        {:defmodule, _meta, [_name, body]} = module, modules when is_list(body) ->
+          {module, [module | modules]}
+
+        node, modules ->
+          {node, modules}
+      end)
+
+    Enum.reverse(modules)
+  end
+
   @doc "Returns true if `node` is inside a loop body (reduce/map/for/recursion)."
   def inside_loop?(node, function) do
     ancestors = ancestors_of(node.id, function)